Create utils apis/getApisFromXExpression
